# wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api Comercial",
    "description": "API para la gestión de documentos comerciales",
    "contact": {
      "name": "Soporte API",
      "email": ""
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Documentos/estado-documentos": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Obtiene el listado de los tipos de documentos disponibles en el sistema.",
        "operationId": "Documentos_ObtenerDocumentos",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoDocumentoResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/insertar-estado-documento": {
      "post": {
        "tags": [
          "Documentos"
        ],
        "summary": "Inserta un nuevo tipo documento en el sistema.",
        "operationId": "Documentos_InsertarDocumento",
        "requestBody": {
          "x-name": "documento",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoDocumentoRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/documentacion-origen": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Obtiene la lista de la documentación vehicular de origen.",
        "operationId": "Documentos_ObtenerListadoDocumentacionOrigen",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentacionOrigenResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/insertar-documentacion-origen": {
      "post": {
        "tags": [
          "Documentos"
        ],
        "summary": "Inserta la documentación de origen del vehículo.",
        "operationId": "Documentos_InsertarDocumentacionOrigen",
        "requestBody": {
          "x-name": "documentacionOrigen",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentacionOrigenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/subir-archivo-documentacion-origen-base64": {
      "post": {
        "tags": [
          "Documentos"
        ],
        "operationId": "Documentos_SubirArchivoBase64",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchivoDocumentoOrigenBase64Request"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/documentos/{documentacionOrigenId}/archivos": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "operationId": "Documentos_ObtenerArchivosPorDocumentacion",
        "parameters": [
          {
            "name": "documentacionOrigenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArchivoDocumentoOrigenResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/insertar-escribania": {
      "post": {
        "tags": [
          "Documentos"
        ],
        "summary": "Inserta una nueva escribanía en el sistema.",
        "operationId": "Documentos_InsertarEscribania",
        "requestBody": {
          "x-name": "escribaniaRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscribaniaRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/escribanias": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Obtiene el listado de escribanías registradas en el sistema.",
        "operationId": "Documentos_ObtenerEscribanias",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EscribaniaResponse"
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/documentacion-post-venta/{documentacionPostVentaId}/archivos": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Obtiene los archivos asociados a una documentación de post venta.",
        "operationId": "Documentos_ObtenerArchivosPorDocumentacionPostVenta",
        "parameters": [
          {
            "name": "documentacionPostVentaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArchivoPostVentaResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Documentos/documentacion-post-venta": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Obtiene la lista de la documentación de post venta.",
        "operationId": "Documentos_ObtenerDocumentacionPostVenta",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentacionPostVentaResponse"
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Clase que muestra los errores",
        "additionalProperties": false,
        "properties": {
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "errorDescripcion": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ErrorType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "error_interno_servidor",
          "validacion_negocio"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "EstadoDocumentoResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "orden": {
            "type": "integer",
            "format": "int32"
          },
          "esFinal": {
            "type": "boolean"
          }
        }
      },
      "EstadoDocumentoRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "orden": {
            "type": "integer",
            "format": "int32"
          },
          "esFinal": {
            "type": "boolean"
          }
        }
      },
      "DocumentacionOrigenResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idChasis": {
            "type": "string",
            "nullable": true
          },
          "fechaRecepcion": {
            "type": "string",
            "format": "date-time"
          },
          "observacion": {
            "type": "string",
            "nullable": true
          },
          "registradoPor": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "DocumentacionOrigenRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idChasis": {
            "type": "string",
            "nullable": true
          },
          "fechaRecepcion": {
            "type": "string",
            "format": "date-time"
          },
          "observacion": {
            "type": "string",
            "nullable": true
          },
          "registradoPor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Ingrese: Bearer {su token JWT}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}